2-add-log-class: Added Log class.
* Added spdlog dependency and updated premake file.
* Replaced most trace message to use the log class.
* Also fixed a warning for texture type and the bugs that happened afterwards.

+++ Renderer/src/Rendering/Material.h
@@ -15,6 +15,6 @@ public:
 	const std::string& GetName() const;
 
 private:
-	Texture textures[Texture::Type::Count];
+	Texture textures[(int)Texture::Type::Count];
 	std::string name;
 };
